agenticai-2/requirements.txt
openai
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq